+++ EC910GroupProject/src/org/com/essex/ec910/artificialstockmarket/strategy/jonathan/SimpleSMATrader.java
@@ -85,7 +85,7 @@ public class SimpleSMATrader extends AbstractTrader {
 					if((outShort[0] > outLong[0]) && portfolio.getShares() == 0)
 					{
 						//Buy as much as possible if short crosses long bottom up 
-						int vol = (int) Math.ceil(this.getInvestableMoney()/spotprice);
+						long vol = (long) Math.ceil(this.getInvestableMoney()/spotprice);
 
 						if(vol > max_buy)
 						{	vol = max_buy;}
